refactor: break the POC into files
